fixup! Show CMS content in sidebar of shuttles page
- Green line controller hijacks shuttle view
- Abstract paragraph retrieval into common helper
- Use helper to assign :paragraphs in each controller
